pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "callreview"
version = "0.1.0"
description = "Internal call recording review system with ingest, transcription pipeline, and web UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Flask>=3.0.0",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]